+++ packages/server/src/http/internals/HTTPHandlerOptions.ts
@@ -1,9 +1,5 @@
 import { Maybe } from '@trpc/server';
-import {
-  BaseHandlerOptions,
-  BaseRequest,
-  BaseResponse,
-} from '../../adapters/node-http/BaseHandlerOptions';
+import { BaseHandlerOptions } from '../../adapters/node-http/BaseHandlerOptions';
 import {
   AnyRouter,
   inferRouterContext,
@@ -12,7 +8,6 @@ import {
 } from '../../router';
 import { TRPCResponse } from '../../rpc';
 import { TRPCError } from '../../TRPCError';
-import { CreateContextFn } from '../resolveHttpResponse';
 import { ResponseMeta } from '../ResponseMeta';
 import { HTTPRequest } from './types';
 
@@ -36,27 +31,6 @@ export interface HTTPHandlerOptionsBase<TRouter extends AnyRouter, TRequest>
    */
   responseMeta?: ResponseMetaFn<TRouter>;
 }
-
-export type HTTPHandlerOptions<
-  TRouter extends AnyRouter,
-  TRequest extends BaseRequest,
-  TResponse extends BaseResponse,
-> = HTTPHandlerOptionsBase<TRouter, TRequest> & {
-  teardown?: () => Promise<void>;
-  maxBodySize?: number;
-} & (inferRouterContext<TRouter> extends void
-    ? {
-        /**
-         * @link https://trpc.io/docs/context
-         **/
-        createContext?: CreateContextFn<TRouter, TRequest, TResponse>;
-      }
-    : {
-        /**
-         * @link https://trpc.io/docs/context
-         **/
-        createContext: CreateContextFn<TRouter, TRequest, TResponse>;
-      });
 
 export interface ResolveHTTPRequestOptions<
   TRouter extends AnyRouter,
